finished character stats. need to get rest.
