+++ FF8/Menu/Kernel_bin.Gender.cs
@@ -1,0 +1,11 @@
+﻿namespace FF8
+{
+    public partial class Kernel_bin
+    {
+        public enum Gender
+        {
+            Male,
+            Female
+        }
+    }
+}
